% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chngp.nonpar.r
\name{chngp.nonpar}
\alias{chngp.nonpar}
\title{A changing point analysis function modified from Qian, S. 's code}
\usage{
chngp.nonpar(data, xvar, yvar, nindex = NULL, num = T, bin = F)
}
\arguments{
\item{data}{A input data frame, default to be NULL,}

\item{xvar}{x variable, either column name or index value; stressor variable}

\item{yvar}{y variable, either column name or index value; response variable}

\item{nindex}{the column to define bins}

\item{num}{boolean variables to determine if Gausian or binomial function}

\item{bin}{binomial}
}
\value{
Returns . 
 sum((x-xmean)^2)) = sum(x^2)- (sum(x))^2/length(x)
}
\description{
This function performed changing point analysis similar to rpart package.However, pvalues are 
computed in this function, as well as confidence intervals.
}
\examples{
set.seed(1)
n = 50
x <- rlnorm(n) + 1:n
x <- (x - min(x))/max(x)
y <- n + 1:n + (rnorm(n))*0.2*n
pvalue <- chngp.nonpar(data, xvar, yvar,nindex=NULL, num = T)
}
\keyword{change}
\keyword{interval,}
\keyword{plot,}
\keyword{point,}
\keyword{regression}
\keyword{scatter}
