% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.CondProb.CI.r
\name{func.CondProb.CI}
\alias{func.CondProb.CI}
\title{This function calculates confidence interval of conditional probability values and plots}
\usage{
func.CondProb.CI(my.data, xindex, yindex, weight = NULL, goodvar = FALSE,
  goodmetric = TRUE, CI = TRUE, j, timeboot = 99,
  xlabel = names(my.data[xindex]), ylabel = names(my.data[yindex]),
  cpplot = FALSE, log.flag = "", change.pt = FALSE, rounder1 = 3,
  rounder2 = 1, sequencecode = "", ...)
}
\arguments{
\item{my.data}{data frame}

\item{xindex}{xvariable at column name or index}

\item{yindex}{xvariable at column name or index}

\item{weight}{is also optional, when there is a site weight column in mydata, enter the column number}

\item{goodvar}{is a flag for good or bad variables, goodvar is 1 if a predictor increase a good metric, e.g., habitat score}

\item{goodmetric}{is a flag for good or bad metrics, e.g., EPT metric is good(1), HBI is bad (0).}

\item{j}{x is stressor, y is biological response, j is the criteria}

\item{cpplot}{is a optional flag to determine if conditional prob should be plotted conditional prob with confidence intervals should be plotted.}

\item{change.pt}{is an option flag to choose if changepoint analysis should be performed}

\item{rounder1}{is number of decimal for change points displaying in graphs}

\item{rounder2}{is biocriteria rounder}
}
\description{
It also uses bootstrapping techniques to calculate confidence interval of change points
}
\examples{

}
