% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multreg.plot.r
\name{multreg.plot}
\alias{multreg.plot}
\title{A regression function created to allow for easy regression related functional performance}
\usage{
multreg.plot(data, yvar = "MBISQ", xvar, j = j, xvar2 = NULL,
  jneg = TRUE, xlim = NULL, x2lim = NULL, binom = TRUE, addcor = F,
  addpred = F, addref = TRUE, ylim = NULL, xlab = names(data[xvar]),
  x2lab = names(data[xvar2]), ln.col = "red",
  ylab = "Probability of Meeting Biological Criterion", cex.lab = 1,
  cex = 1, cex.axis = 1, las = 1, pch = 21, col = 1, bg = 8,
  main = "", type = "p", cex.main = 1, varlist = NULL, varpred,
  cont = TRUE, cont2 = FALSE, xlog = TRUE, x2log = TRUE, add = F,
  rounder = 2, nbin = 41, LC50.plot = FALSE, model.sel = NULL, ...)
}
\arguments{
\item{data}{A input data frame, default to be NULL,}

\item{yvar}{for column index of x, y values in data data frame, or column names; y could be binary}

\item{xvar}{x variable, either column name or index value}

\item{j}{threshold for logistic regression}

\item{binom}{if binomial function will be used}

\item{xlog}{if xvar should be log transformed}

\item{rounder}{if x should be round to a value}

\item{model.sel}{= 0 (single var) 1 (2 var), 2 (2 var interaction),3(quadratic)}

\item{x.add}{1 to convert percent x value = 0 to 1+ x value , or 100, then *100+1,}

\item{x2var}{x variable, either column name or index value}

\item{xlog2}{if xvar2 should be log transformed}

\item{add:}{add multiple regression lines in single graphs espeically for logistic, regression single x variable}

\item{addcor:}{add principal component line between two explanatory variables and the predicted line for xvar}

\item{addref:}{add reference value for one variable to predict the other, e.g, add reference value for habitat(160) to predict cond}

\item{nbin:}{logistic regression number of bins.}

\item{LC50.plot:}{to add the biological response variable plot}

\item{varlist:}{additional variables to include in the multiple regression plot, all these variables should be normal distributed}

\item{varpred:}{additional variable values for making predictions}

\item{jneg:}{xvar and yvar are negatively correlated, default is true}

\item{cont:}{if true plot contour, if not plot scatterplot}

\item{cont2:}{if true plot contour, if not plot scatterplot}
}
\value{
Returns
}
\description{
A regression function created to allow for easy regression related functional performance
}
\examples{
set.seed(1)
x <- rlnorm(n) + 1:n
x <- (x - min(x))/max(x)
y <- n + 1:n + (rnorm(n))*0.2*n
par(mfrow=c(1,2))
plot(y~x)
scatter.plot(data = data.frame(x,y), xvar= "x", yvar = "y", add.fit = "linear", xlim = c(0,n), 
    ylim = range(y), log.flag = "x", x.add =100)
}
\keyword{logistic}
\keyword{multiple}
\keyword{regression}
\keyword{regression,}

