% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel_data}
\alias{panel_data}
\title{Panel Data}
\format{A data frame with 50,000 rows and four variables:
\describe{
    \item{firmid}{A unique firm identifier.}
    \item{year}{The year of the observation.}
    \item{y}{The dependent variable.}
    \item{x}{The independent variable.}
}}
\source{
Find the data \href{http://www.kellogg.northwestern.edu/faculty/petersen/htm/papers/se/test_data.htm}{here}.
}
\usage{
panel_data
}
\description{
A dataset containing a toy example of panel data.
}
\keyword{datasets}
