% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.R
\name{reg}
\alias{reg}
\title{Replicate STATA Regressions}
\usage{
reg(formula, data, robust = FALSE, cluster = NULL)
}
\arguments{
\item{formula}{A formula providing the specifications for your model.}

\item{data}{A data frame on which to run your regression.}

\item{robust}{A logical value to indicate if the model should use robust
standard errors.}

\item{cluster}{A character vector to specify up to two variables by which
to cluster standard errors.}
}
\value{
A data frame with one row for each independent variable and columns
  containing the estimated coefficients and the associated standard errors,
  t-statistics, and p-values.
}
\description{
\code{reg} replicates STATA regressions including regressions with robust standard
errors and clustered standard errors.
}
\examples{
reg(y ~ x, my_data, robust = TRUE)
reg(y ~ x, my_data, cluster = "variable")
reg(y ~ x, my_data, cluster = c("variable_1", "variable_2"))
}
