% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.plot.r
\name{scatter.plot}
\alias{scatter.plot}
\title{A regression function for simple regression and break point analysis and plotting}
\usage{
scatter.plot(data = NULL, xvar, yvar, xName = "x", yName = "y",
  xlim = NULL, ylim = NULL, sign = FALSE, xlab = names(data[xvar]),
  ylab = names(data[yvar]), addaxes = TRUE, cex.eq = 1.5, bty = "o",
  type = "p", pch = 1, col = 1, main = "", cex.main = 1,
  cex.lab = 1, cex = 1, cex.axis = 1, lty = 1, las = 1, bg = "gray",
  side = 4, tx.col = 2, xlog.scale = 10, ylog.scale = 10,
  lab.pos = "topleft", labs = "", x.add = 0, add.fit = "none",
  interval = "none", ln.col = "red", level = 0.95, add.cor = FALSE,
  add.reg = FALSE, add.rq = FALSE, nl.rq = FALSE, rq.tau = 0.5,
  log = "", add = FALSE, add.r2 = FALSE, add.interp = FALSE,
  pred.x = NA, pred.y = NA, rounder1 = 2, rounder2 = 0, ...)
}
\arguments{
\item{data}{A input data frame, default to be NULL,}

\item{xvar}{x variable, either column name or index value}

\item{yvar}{y variable, either column name or index value}

\item{xName}{xName for making equation labels}

\item{yName}{yName for making equation labels}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{sign}{if TRUE then use signif for equation otherwise, use round}

\item{xlab}{xlabel}

\item{ylab}{ylabel}

\item{addaxes}{to plot axes}

\item{bty}{add a box frame}

\item{pch}{point type}

\item{cex}{magnifier}

\item{cex.axis}{# magnifier}

\item{lty}{line type}

\item{las}{direct}

\item{bg}{background col}

\item{side}{which side to add}

\item{tx.col}{is for col of labs ln.col for fitted line color}

\item{xlog.scale}{either 10 or exp}

\item{ylog.scale}{either 10 or exp}

\item{lab.pos}{control position of labels "bottomright" "bottomleft", "topright", "topleft"}

\item{labs}{add labels}

\item{x.add}{1 to convert percent x value = 0 to 1+ x value , or 100, then *100+1,}

\item{add.fit}{either "linear", "lowess", "loess", "segment" (segmented regression), "change" (regression tree)}

\item{level}{CI level}

\item{add.cor}{add spearman correlation,}

\item{add.reg}{will add regression model}

\item{add.rq}{# add regression quantile}

\item{rq.tau}{a vector to indicate which quantiles we want to use}

\item{log}{will control x,y axis if log,}

\item{add}{if TRUE, then add to existing graph}

\item{add.r2}{a boolean to indicate if r2}

\item{add.interp}{will add a interpolation for inverse prediction of x based on pred.y}

\item{pred.x}{used to predict x value}

\item{pred.y}{used to predict y}

\item{rounder1}{round number of digits for x value}

\item{rounder2}{round number of digits for y value}

\item{yvar}{for column index of x, y values in data data frame, or column names}

\item{inteval}{"confidence", "prediction", "none"}

\item{rq}{choose if we want to do quantile regression}

\item{nlrq}{only works for logistic fit}

\item{addaxes}{if true, add axes}

\item{ex.eq}{equation}

\item{x.add}{add x}
}
\value{
Returns . 
 sum((x-xmean)^2)) = sum(x^2)- (sum(x))^2/length(x)
}
\description{
This function performed basic regression analysis with the option to add confidence intervals and
prediction intervals; alternative regression approaches, e.g., segmented regression, change point
analysis (through rpart) and confidence bounds, quantile regression, and lowess regression splines.
This version also added a capacity for transforming percent values (where 0 is probably the minimum)
by adding 1 or 100 before log transformation if users choose so. 
Any x variable with 0s and log is true, the add 1 to the original or 100 times of the original.
}
\examples{
set.seed(1)
n = 50
x <- rlnorm(n) + 1:n
x <- (x - min(x))/max(x)
y <- n + 1:n + (rnorm(n))*0.2*n
par(mfrow=c(1,2))
plot(y~x)
scatter.plot(data = data.frame(x,y), xvar= "x", yvar = "y", add.fit = "linear", xlim = c(0,n), 
    ylim = range(y), log = "x", x.add =100)
}
\keyword{change}
\keyword{interval,}
\keyword{plot,}
\keyword{point,}
\keyword{regression}
\keyword{scatter}
