% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf3d.r
\name{surf3d}
\alias{surf3d}
\title{A three dimensional plot function}
\usage{
surf3d(x, y, z, xlab = "X Coordinate (feet)", ylab = "Y Coordinate (yyy)",
  border = NA, zlab = "Elevation", main = "Surface elevation data",
  phi = 45, theta = 45, ncut = 100, scheme = c("purple4", "blue4",
  "blue1", "green3", "yellowgreen", "yellow2", "sandybrown", "red", "darkred"),
  type = "rgb")
}
\arguments{
\item{x}{}

\item{y}{a boolean variable deciding if the box should be horizontal}

\item{z}{e if an outline}

\item{border}{border of lines}

\item{phi}{}

\item{theta}{The function arguments phi and theta are used to rotate the viewing angle of the surface}

\item{box}{should hte bounding box for the surface be displayed}
}
\description{
The function persp is the base graphics function for creating wireframe surface plots.
adopted from http://www.r-bloggers.com/creating-surface-plots/
The persp function requires a list of x and y values covering the grid of vertical values
 which is specified as the z variable. The heights for the display are specified as a table
 of values which we saved previously as the object z during the calculations when the local
 trend surface model was fitted to the data.
}
