% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeaxes.r
\name{threeaxes.plot}
\alias{threeaxes.plot}
\title{A function to plot 2 y axes}
\usage{
threeaxes.plot(my.data = NULL, xvar, yvar1, yvar2,
  xlab = names(my.data[xvar]), y1lab = names(my.data[yvar1]),
  y2lab = names(my.data[yvar1]), pch1 = 21, col1 = 4, bg1 = "gray",
  pch2 = 22, col2 = 2, bg2 = "gray", main = "", cex.main = 1,
  cex.lab = 1, cex = 1, cex.axis = 1, type = "p", las1 = 1,
  las2 = 1, xlog = FALSE, y1log = FALSE, mar = c(5, 4, 3, 4),
  y2log = FALSE, eq.scale = TRUE, x0lim = NULL, y1lim = NULL,
  y2lim = NULL, ...)
}
\arguments{
\item{xvar}{x variable, either column name or index value}

\item{yvar1}{y variable, either column name or index value}

\item{yvar2}{y2 variable, either column name or index value}

\item{xlab}{axis 1 name}

\item{y1lab}{axis 2 name}

\item{y2lab}{axis 4 name}

\item{pch1}{yvar1 sympbol}

\item{col1}{yvar1 color}

\item{bg1}{yvar1 symbol background}

\item{pch2}{yvar2 sympbol}

\item{col2}{yvar2 color}

\item{bg2}{yvar2 symbol background}

\item{y1log}{if true, yvar1 should be log transformed}

\item{y2log}{if true, yvar2 should be log transformed}

\item{eq.scale}{=T then y1 and y2 have same number of ticks and labels, if false, y2 scale itself.}

\item{x0lim}{add extra values to make sure xlim fall into the self defined range}

\item{y1lim}{add extra values to make sure y1lim fall into the self defined range}

\item{y2lim}{add extra values to make sure y2lim fall into the self defined range}

\item{data}{A input data frame, default to be NULL,}

\item{use}{is.list(x) to check if x variable is a time object...}
}
\value{
Returns .
}
\description{
This function plot two response variables vs. same x axis (predictors)
}
\examples{
xvar =1:100; yvar1 = rnorm(100)+ 1:100; yvar2 = rlnorm(100) + 1:100 
par(mfrow=c(1,3))
threeaxes.plot(my.data = NULL, xvar, yvar1, yvar2, y2log =T, eq.scale =F) 
threeaxes.plot(my.data = NULL, xvar, yvar1, yvar2, y2log =T, eq.scale =T) 
threeaxes.plot(my.data = NULL, xvar, yvar1, yvar2, y2log =F, eq.scale =F) 
}
\keyword{axes}
\keyword{respnse}
\keyword{three}
\keyword{varialbes,}
